Add Migration Script and fix Dataset Table Permissions

+++ backend/dataall/db/api/dataset.py
@@ -371,11 +371,13 @@ class Dataset:
 
     @staticmethod
     def transfer_stewardship_to_new_stewards(session, dataset, new_stewards):
-        ResourcePolicy.delete_resource_policy(
-            session=session,
-            group=dataset.stewards,
-            resource_uri=dataset.datasetUri,
-        )
+        env = Environment.get_environment_by_uri(session, dataset.environmentUri)
+        if dataset.stewards != env.SamlGroupName:
+            ResourcePolicy.delete_resource_policy(
+                session=session,
+                group=dataset.stewards,
+                resource_uri=dataset.datasetUri,
+            )
         ResourcePolicy.attach_resource_policy(
             session=session,
             group=new_stewards,
@@ -383,13 +385,15 @@ class Dataset:
             resource_uri=dataset.datasetUri,
             resource_type=models.Dataset.__name__,
         )
+
         dataset_tables = [t.tableUri for t in Dataset.get_dataset_tables(session, dataset.datasetUri)]
         for tableUri in dataset_tables:
-            ResourcePolicy.delete_resource_policy(
-                session=session,
-                group=dataset.stewards,
-                resource_uri=tableUri,
-            )
+            if dataset.stewards != env.SamlGroupName:
+                ResourcePolicy.delete_resource_policy(
+                    session=session,
+                    group=dataset.stewards,
+                    resource_uri=tableUri,
+                )
             ResourcePolicy.attach_resource_policy(
                 session=session,
                 group=new_stewards,
@@ -415,11 +419,6 @@ class Dataset:
                 ResourcePolicy.delete_resource_policy(
                     session=session,
                     group=dataset.stewards,
-                    resource_uri=share.shareUri,
-                )
-                ResourcePolicy.delete_resource_policy(
-                    session=session,
-                    group=dataset.SamlAdminGroupName,
                     resource_uri=share.shareUri,
                 )
         return dataset
@@ -542,6 +541,11 @@ class Dataset:
         ResourcePolicy.delete_resource_policy(
             session=session, resource_uri=uri, group=dataset.SamlAdminGroupName
         )
+        env = Environment.get_environment_by_uri(session, dataset.environmentUri)
+        if dataset.SamlAdminGroupName != env.SamlGroupName:
+            ResourcePolicy.delete_resource_policy(
+                session=session, resource_uri=uri, group=env.SamlGroupName
+            )
         if dataset.stewards:
             ResourcePolicy.delete_resource_policy(
                 session=session, resource_uri=uri, group=dataset.stewards
